requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.8
cssselect==1.2.0
decorator==4.4.2
ensta==5.2.9
fake-useragent==1.5.1
future==1.0.0
h11==0.14.0
html2image==2.0.4.3
huepy==1.2.1
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
imgkit==1.2.3
importlib_metadata==8.0.0
iniconfig==2.0.0
instagrapi==2.1.2
lxml==5.2.2
lxml_html_clean==0.1.1
mock==5.1.0
moviepy==1.0.3
ntplib==0.4.0
numpy==2.0.0
outcome==1.3.0.post0
packaging==24.1
parse==1.20.2
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.10
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
pyotp==2.9.0
pyppeteer==2.0.0
pyquery==2.0.0
PySocks==1.7.1
pytest==8.2.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-html==0.10.0
requests-toolbelt==1.0.0
responses==0.25.3
schedule==1.2.2
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.19
w3lib==2.2.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.19.2
